Move some code to a new c_analyzer.variables package.
