+++ Lib/test/test_tools/test_c_analyzer/test_c_globals/test_supported.py
@@ -4,8 +4,11 @@ import unittest
 
 from .. import tool_imports_for_tests
 with tool_imports_for_tests():
-    from c_analyzer.common.info import ID, Variable
-    from c_globals.supported import is_supported, ignored_from_file
+    from c_analyzer.common.info import ID
+    from c_analyzer.variables.info import Variable
+    from c_globals.supported import (
+            is_supported, ignored_from_file,
+            )
 
 
 class IsSupportedTests(unittest.TestCase):
