+++ Lib/test/test_tools/test_c_analyzer/test_variables/__init__.py
@@ -1,0 +1,6 @@
+import os.path
+from test.support import load_package_tests
+
+
+def load_tests(*args):
+    return load_package_tests(os.path.dirname(__file__), *args)
